Changed type of Sum fold from int to long to allow for ome larger tree computation

+++ apgas.glb/src/apgas/glb/example/Sum.java
@@ -9,7 +9,7 @@ import apgas.glb.Fold;
 
 /**
  * Performs the addition on integers.
- * 
+ *
  * @author Patrick Finnerty
  *
  */
@@ -21,7 +21,7 @@ public class Sum implements Fold<Sum>, Serializable {
   private static final long serialVersionUID = 3582168956043482749L;
 
   /** Integer in which the sum is performed */
-  public int sum;
+  public long sum;
 
   /*
    * (non-Javadoc)
